services/ComfyUI-Filen-Sync/requirements.txt
requests
watchdog
python-gnupg